+++ src/main/java/org/jhipster/web/rest/EntryResource.java
@@ -56,7 +56,8 @@ public class EntryResource {
         if (entry.getId() != null) {
             throw new BadRequestAlertException("A new entry cannot already have an ID", ENTITY_NAME, "idexists");
         }
-        if (!entry.getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
+        if (entry.getBlog() != null &&
+            !entry.getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
             return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
         }
         Entry result = entryRepository.save(entry);
@@ -78,7 +79,8 @@ public class EntryResource {
     @Timed
     public ResponseEntity<?> updateEntry(@Valid @RequestBody Entry entry) throws URISyntaxException {
         log.debug("REST request to update Entry : {}", entry);
-        if (!entry.getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
+        if (entry.getBlog() != null &&
+            !entry.getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
             return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
         }
         if (entry.getId() == null) {
@@ -116,10 +118,11 @@ public class EntryResource {
     public ResponseEntity<?> getEntry(@PathVariable Long id) {
         log.debug("REST request to get Entry : {}", id);
         Entry entry = entryRepository.findOneWithEagerRelationships(id);
-        if (!entry.getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
+        if (entry != null && entry.getBlog() != null &&
+            !entry.getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
             return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
         }
-        return ResponseUtil.wrapOrNotFound(Optional.of(entry));
+        return ResponseUtil.wrapOrNotFound(Optional.ofNullable(entry));
     }
 
     /**
@@ -130,8 +133,13 @@ public class EntryResource {
      */
     @DeleteMapping("/entries/{id}")
     @Timed
-    public ResponseEntity<Void> deleteEntry(@PathVariable Long id) {
+    public ResponseEntity<?> deleteEntry(@PathVariable Long id) {
         log.debug("REST request to delete Entry : {}", id);
+        Entry entry = entryRepository.findOneWithEagerRelationships(id);
+        if (entry != null && entry.getBlog() != null &&
+            !entry.getBlog().getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(""))) {
+            return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
+        }
         entryRepository.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
     }
